+++ src/main/routes/features/queryManagement/qmViewQueriesController.ts
@@ -8,7 +8,8 @@ import { AppRequest } from 'models/AppRequest';
 import config from 'config';
 import { CivilServiceClient } from 'client/civilServiceClient';
 import {constructResponseUrlWithIdParams} from 'common/utils/urlFormatter';
-import {ViewObjects, ViewQueriesService} from 'services/features/queryManagement/viewQueriesService';
+import {ViewQueriesService} from 'services/features/queryManagement/viewQueriesService';
+import {ViewObjects} from "form/models/queryManagement/viewQuery";
 
 const civilServiceApiBaseUrl = config.get<string>('services.civilService.url');
 const civilServiceClient: CivilServiceClient = new CivilServiceClient(civilServiceApiBaseUrl);
